requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
Django==4.2.15
djangorestframework==3.15.2
h11==0.14.0
idna==3.7
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
